feat(tui): selection & scroll
